implement invoices logic

+++ app/Http/Controllers/WorkController.php
@@ -43,8 +43,15 @@ class WorkController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Work $work): View
+    public function edit(Work $work)
     {
+        if ($work->invoice()->count()) {
+            return Redirect::route('loads.index')->with('flash', [
+                'status' => 'warning',
+                'text' => 'Work has posted invoice. Work can\'t be edited.' ,
+            ]);
+        }
+
         return view('work.edit', [
             'work' => $work,
             'drivers' => Driver::all()->sortBy('first_name'),
@@ -56,6 +63,13 @@ class WorkController extends Controller
      */
     public function update(UpdateWorkRequest $request, Work $work): RedirectResponse
     {
+        if ($work->invoice()->count()) {
+            return Redirect::back()->with('flash', [
+                'status' => 'fail',
+                'text' => 'Work has posted invoice. Work can\'t be updated.' ,
+            ]);
+        }
+
         $flash = ['status' => 'success', 'text' => 'Work data updated.'];
 
         if (!$work->update($request->validated())) {
@@ -70,7 +84,13 @@ class WorkController extends Controller
      */
     public function destroy(Work $work): RedirectResponse
     {
-        // cant be deleted if invoiced
+        if ($work->invoice()->count()) {
+            return Redirect::back()->with('flash', [
+                'status' => 'fail',
+                'text' => 'Work has related invoice. Work can\'t be deleted.' ,
+            ]);
+        }
+
         $work->delete();
 
         return Redirect::route('loads.index')->with('flash', ['status' => 'success', 'text' => 'Work deleted.']);
